Eliminate layout shift in bell hover
The bell has a responsive effect on hover when the user is subscribed.
This was implemented by adding or removing the "small" css class for the
bell.  The small/medium/large classes are also used to expose size
options to customers.

This change transitions to an approach where we add a scale transform
CSS property, which does not re-trigger layout.